Reworking for basic validator output
