+++ libmei/tools/config.yml
@@ -5,3 +5,4 @@ addons_dir: "../addons"
 elements: false
 namespace: "vrv"
 datatypes: "./datatypes.yml"
+basic_odd: "../mei/develop/mei-basic_compiled.odd"
